requirements.txt
fastapi>=0.100.0
uvicorn>=0.22.0
llama-index-core>=0.10.0
llama-index-llms-openai>=0.1.0
openai>=1.0.0
pydantic>=2.0
pdfplumber>=0.10.0
pdf2image>=1.16.3
pytesseract>=0.3.10
pillow>=10.0.0